Fix minor

+++ Examples/Source/TestRendering.h
@@ -62,7 +62,7 @@ inline auto TestRendering()
 		auto TopLight = world.SpawnActor<PointLightActor>("TopLight");
 		auto BottomLight = world.SpawnActor<PointLightActor>("BottomLight");
 
-		world.TickFunction = [&, TopLight, BottomLight](float DeltaTime, World& world) {
+		world.TickFunction = [&, TopLight, BottomLight](double DeltaTime, World& world) {
 			if(LightMotion) {
 				static float TotalTime = 0;
 				TotalTime += DeltaTime * 2.f;
